\name{SLMSeg-package}
\alias{SLMSeg-package}
\alias{SLMSeg}
\docType{package}
\title{
\packageTitle{SLMSeg}
}
\description{
\packageDescription{SLMSeg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SLMSeg}
\packageIndices{SLMSeg}
%% ~~ An overview of how to use the package, including the most ~~
%% ~~ important functions ~~
}
\author{
\packageAuthor{SLMSeg}

Maintainer: \packageMaintainer{SLMSeg}
}
\references{
%%~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
%% ~~ Optional links to other man pages, e.g. ~~
%% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
